+++ src/game/tests.rs
@@ -332,6 +332,7 @@ mod vintage_cube_brainsurge;
 mod vintage_cube_brazen_borrower;
 mod vintage_cube_breach;
 mod vintage_cube_brightglass_gearhulk;
+mod vintage_cube_broadside_bombardiers;
 mod vintage_cube_cankerbloom;
 mod vintage_cube_carnage_interpreter;
 mod vintage_cube_caryatid;